+++ tuxemon/log.py
@@ -12,9 +12,24 @@ from tuxemon import prepare
 from tuxemon.constants import paths
 
 
+def get_git_hash() -> str:
+    """Gets the current Git hash."""
+    try:
+        return (
+            subprocess.check_output(["git", "describe", "--always"])
+            .strip()
+            .decode()
+        )
+    except subprocess.CalledProcessError:
+        logging.warning("Git command failed. Git hash not available.")
+        return "N/A"
+    except FileNotFoundError:
+        logging.warning("Git not found. Git hash not available.")
+        return "N/A"
+
+
 def configure() -> None:
-    """Configure logging based on the settings in the config file."""
-    # Set our logging levels
+    """Configures logging based on the settings in the config file."""
     LOG_LEVELS = {
         "debug": logging.DEBUG,
         "info": logging.INFO,
@@ -22,40 +37,25 @@ def configure() -> None:
         "error": logging.ERROR,
         "critical": logging.CRITICAL,
     }
-    config = prepare.CONFIG
+    config = prepare.CONFIG.logging
     loggers = {}
 
-    if config.debug_level in LOG_LEVELS:
-        log_level = LOG_LEVELS[config.debug_level]
-    else:
-        log_level = logging.INFO
+    log_level = LOG_LEVELS.get(config.debug_level, logging.INFO)
 
-    # Set up logging if the configuration has it enabled
     if config.debug_logging:
-        # Enable suppressed warnings
         warnings.filterwarnings("default")
 
-        for logger_name in config.loggers:
-            # Get the current git hash
-            try:
-                githash = (
-                    subprocess.check_output(["git", "describe", "--always"])
-                    .strip()
-                    .decode()
-                )
-                print(f"Git Hash: {githash}")
-            except:
-                print("No Git Hash")
+        githash = get_git_hash()
+        print(f"Git Hash: {githash}")
 
-            # Enable logging for all modules if specified.
+        for logger_name in config.loggers:
             if logger_name == "all":
                 print("Enabling logging of all modules.")
                 logger = logging.getLogger()
             else:
-                print("Enabling logging for module: %s" % logger_name)
+                print(f"Enabling logging for module: {logger_name}")
                 logger = logging.getLogger(logger_name)
 
-            # Enable logging
             log_formatter = logging.Formatter(
                 "[%(asctime)s] %(name)s - %(levelname)s - %(message)s"
             )
@@ -65,38 +65,38 @@ def configure() -> None:
             log_strm.setFormatter(log_formatter)
             logger.addHandler(log_strm)
 
-            # Enable logging to file
             if config.log_to_file:
                 log_dir = os.path.realpath(f"{paths.USER_STORAGE_DIR}/logs")
-                if not os.path.exists(log_dir):
-                    os.makedirs(log_dir)
+                os.makedirs(
+                    log_dir, exist_ok=True
+                )  # creates the directory if it does not exists.
                 if config.log_keep_max > 0:
-                    log_dir_files = {}
-                    for entry in os.listdir(log_dir):
-                        log_dir_files[entry] = os.stat(
-                            f"{log_dir}/{entry}"
-                        ).st_mtime
+                    log_dir_files = {
+                        entry.name: entry.stat().st_mtime
+                        for entry in os.scandir(log_dir)
+                        if entry.is_file()
+                    }
                     sorted_files = sorted(
                         log_dir_files.items(), key=itemgetter(1), reverse=True
                     )
-                    log_dir_files.clear()
+
                     if len(sorted_files) > config.log_keep_max:
-                        for x in range(
-                            config.log_keep_max - 1, len(sorted_files)
-                        ):
-                            os.remove(f"{log_dir}/{sorted_files[x][0]}")
-                formatted_time = time.strftime(
-                    "%Y-%m-%d_%Hh%Mm%Ss", time.localtime()
-                )
-                log_file = logging.FileHandler(
-                    f"{log_dir}/{formatted_time}.log"
-                )
-                log_file.setFormatter(log_formatter)
-                log_file.setLevel(log_level)
-                logger.addHandler(log_file)
+                        for x in range(config.log_keep_max, len(sorted_files)):
+                            filename = sorted_files[x][0]
+                            file_path = os.path.join(log_dir, filename)
+                            os.remove(file_path)
+
+                    formatted_time = time.strftime(
+                        "%Y-%m-%d_%Hh%Mm%Ss", time.localtime()
+                    )
+                    log_file = logging.FileHandler(
+                        os.path.join(log_dir, f"{formatted_time}.log")
+                    )
+                    log_file.setFormatter(log_formatter)
+                    log_file.setLevel(log_level)
+                    logger.addHandler(log_file)
 
             loggers[logger_name] = logger
 
-            # prevent pyscroll redraw warnings
-            pyscroll_logger = logging.getLogger("orthographic")
-            pyscroll_logger.setLevel(logging.ERROR)
+        pyscroll_logger = logging.getLogger("orthographic")
+        pyscroll_logger.setLevel(logging.ERROR)
