+++ tuxemon/prepare.py
@@ -432,7 +432,7 @@ def pygame_init() -> None:
     SCREEN_RECT = SCREEN.get_rect()
 
     # Disable the mouse cursor visibility
-    pg.mouse.set_visible(not CONFIG.hide_mouse)
+    pg.mouse.set_visible(not CONFIG.controller.hide_mouse)
 
     # Set up any gamepads that we detect
     # The following event types will be generated by the joysticks:
